% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{allProjectData}
\alias{allProjectData}
\title{Project Contribution Data across all chains}
\usage{
allProjectData()
}
\value{
Project Contribution Data
}
\description{
Project Contribution Data across all chains
}
\examples{
\dontrun{
allProjectData()
}
}
