% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{allVoteData}
\alias{allVoteData}
\title{Vote Contribution Data for a give chain}
\usage{
allVoteData()
}
\value{
Vote Contribution Data
}
\description{
Vote Contribution Data for a give chain
}
\examples{
\dontrun{
allVoteData()
}
}
