% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{chainList}
\alias{chainList}
\title{Get a List of supported Chains}
\usage{
chainList()
}
\value{
List of Supported chain Indexes
}
\description{
Get a List of supported Chains
}
\examples{
\dontrun{
chainList()
}
}
