% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{chainProjectData}
\alias{chainProjectData}
\title{Project Contribution Data for a give chain}
\usage{
chainProjectData(cidx)
}
\arguments{
\item{cidx}{Chain Index out of 1, 10, 137, 250, 42161, 421613, 424}
}
\value{
Project Contribution Data
}
\description{
Project Contribution Data for a give chain
}
\examples{
\dontrun{
chainProjectData(42161)
}
}
