% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{chainProjectDF}
\alias{chainProjectDF}
\title{Get Projects DataFrame for a given chain index}
\usage{
chainProjectDF(cidx)
}
\arguments{
\item{cidx}{Chain Index out of 1, 10, 137, 250, 42161, 421613, 424}
}
\value{
Project DataFrame
}
\description{
Get Projects DataFrame for a given chain index
}
\examples{
\dontrun{
chainProjectDF(42161)
}
}
