% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{chainRoundDF}
\alias{chainRoundDF}
\title{Get Rounds DataFrame for a given chain index}
\usage{
chainRoundDF(cidx)
}
\arguments{
\item{cidx}{Chain Index out of 1, 10, 137, 250, 42161, 421613, 424}
}
\value{
Rounds DataFrame
}
\description{
Get Rounds DataFrame for a given chain index
}
\examples{
\dontrun{
chainRoundDF(42161)
}
}
