% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{chainVoteData}
\alias{chainVoteData}
\title{Vote Contribution Data for a give chain}
\usage{
chainVoteData(cidx)
}
\arguments{
\item{cidx}{Chain Index out of 1, 10, 137, 250, 42161, 421613, 424}
}
\value{
Vote Contribution Data
}
\description{
Vote Contribution Data for a give chain
}
\examples{
\dontrun{
chainVoteData(42161)
}
}
