% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{projectDF}
\alias{projectDF}
\title{Get Project DataFrame across all chains}
\usage{
projectDF()
}
\value{
Project DataFrame across All Chains
}
\description{
Get Project DataFrame across all chains
}
\examples{
\dontrun{
projectDF()
}
}
