% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{roundDF}
\alias{roundDF}
\title{Get Rounds DataFrame across all chains}
\usage{
roundDF()
}
\value{
Round DataFrame across All Chains
}
\description{
Get Rounds DataFrame across all chains
}
\examples{
\dontrun{
roundDF()
}
}
